moved breader, deepcopy, key and stringfs into exp. moved draw and x11 into extern. made canvas and hello compile. new9p/seq: factorised replier function.

+++ canvas/ellipse.go
@@ -19,7 +19,7 @@ type Ellipse struct {
 }
 
 //color, center, radius a radius b and width
-func NewEllipse(col image.Image, cr image.Point, ra, rb int, width float) *Ellipse {
+func NewEllipse(col image.Image, cr image.Point, ra, rb int, width float64) *Ellipse {
 	obj := new(Ellipse)
 	obj.cr = pixel2fixPoint(cr)
 	obj.ra = int2fix(ra)
